Fix the installation for FreeBSD.
For FreeBSD the standard version for python if you install it with `pkg install python` is version 3.7.
If you now try to enforce version 2.7 the installation will fail as this binary does not exist.
I tested it and installation on node depedencies are working fine with python 3.7.
So the gyp problems are fixed already upstream.

I will create another CR to fix the shebang for the two scripts, so I increase the version here not.