Create versions.json (#510)
* Create versions.json

This file is a replacement for the version part of the depricated file ioBroker.info/master/admin/lib/data/infoData.json.
It will be used by the next version of the windows installer and maybe it makes sense to use it also for the Linux installation script and for setting the state "system.host.<hostname>.versions.npmNewestNext"

* Update versions.json

Renamed node... to nodeJs...

* Update versions.json

Removed windows installer version. The installer will use package.json of iobroker.build to get current installer version.

* Update versions.json

After discussion with Ingo: Updated to currently supported versions.